backend/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
deepgram-sdk==3.8.0
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.7
fpdf==1.7.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
marshmallow==3.26.0
multidict==6.1.0
mypy-extensions==1.0.0
openai==1.60.0
packaging==24.2
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.45.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==13.1
whisper==1.1.10
yarl==1.18.3
